Fixes issue #37
As per hassleg "It's not clear to me what is being asked for in the HOF exercise. A reference is made to a function named add that doesn't exist; I assume this is a typo and that it should be referring to the function named add1. The bigger question is what is meant by the opposite value of that returned by add1? Minus 1?"